@@ -20,26 +20,26 @@
         }
     ],
     "require": {
-        "php": "^7.2 | 8.0.* | 8.1.*",
+        "php": "^7.4 | 8.0.* | 8.1.*",
         "ext-intl": "*",
         "lib-icu": ">=4.8",
         "jms/parser-lib": "^1.0.0",
         "psr/log": "^1.0",
-        "symfony/config": "^3.4.31|^4.0",
-        "symfony/dependency-injection": "^3.4.31|^4.0",
-        "symfony/finder": "^3.4.31|^4.0",
-        "symfony/framework-bundle": "^4.4",
-        "symfony/http-kernel": "^3.4.31|^4.0",
-        "symfony/translation": "^4.2",
-        "symfony/translation-contracts": "^1.0.2|^2.0|^3.0",
+        "symfony/config": "^4.3|^5.0",
+        "symfony/dependency-injection": "^4.4|^5.0",
+        "symfony/finder": "^3.4.31|^4.0|^5.0",
+        "symfony/framework-bundle": "^4.4|^5.0",
+        "symfony/http-kernel": "^4.4|^5.0",
+        "symfony/translation": "^4.2|^5.0",
+        "symfony/translation-contracts": "^2.0|^3.0",
         "twig/twig": "^1.42|^2.0|^3.0"
     },
     "require-dev": {
         "phpunit/phpunit": "^8.5 | ^9.0",
-        "symfony/monolog-bundle": "^3.4.31|^4.0",
-        "symfony/phpunit-bridge": "> 5.0",
-        "symfony/twig-bundle": "^3.4.31|^4.0",
-        "symfony/yaml": "^4.4"
+        "symfony/monolog-bundle": "^3.4.31",
+        "symfony/phpunit-bridge": ">= 5.0",
+        "symfony/twig-bundle": "^3.4.31|^4.0|^5.0",
+        "symfony/yaml": "^4.4|^5.0"
     },
     "autoload": {
         "psr-4": {
